requirements.txt
langchain
langchain-community
langchain-google-genai
chromadb
pymupdf
python-dotenv
langchain-chroma
chainlit